requirements.txt
pdfplumber==0.11.0
camelot-py[cv]==0.11.0
opencv-python==4.8.1.78
